+++ CHANGELOG.md
@@ -43,10 +43,14 @@ to the project during `#hactoberfest`. List of awesome people:
 - Fixes that `MultipleIfsInComprehensionViolation` was not enabled
 - Fixes flaky behaviour of `test_module_names` test package
 - Fixed `TooManyMethodsViolation` not displaying line number in output
+- Fixed `OffsetVisitor` due to python [bug](https://bugs.python.org/issue29205)
 
 ### Misc
 
+- Updates `poetry` version
 - Refactoring: some general changes, including better names and APIs
+- Improves docs: now we have `versionadded` for each violation
+- Improves tests: now we are testing options
 - Improves tests: now we have different `tests/` folder structure
 - Improves tests: now we are testing presets
 - Improves tests: now we are using different logic inside `assert_errors`